pyproject.toml
[project]
name = "friday-tony-stark-demo"
version = "0.1.0"
description = "Friday — A Tony Stark-inspired MCP server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "httpx",
    "livekit-agents[deepgram,groq,openai,sarvam,silero]>=1.5.1",
    "livekit-plugins-google>=1.5.1",
    "psutil>=5.9",
    "python-dotenv",
    # fullvoice mode (mic -> STT -> LLM -> TTS) — see agent_friday.py `fullvoice`.
    # tiny.en model auto-downloads to HF cache on first use (~75MB).
    "faster-whisper>=1.0",
    "sounddevice>=0.4",
    "numpy>=1.24",
]
# Note: `agent_friday.py voice` uses Windows built-in SAPI via a PowerShell
# subprocess (System.Speech.Synthesis).  No extra Python dep required.

[project.scripts]
friday = "server:main"
friday_voice = "agent_friday:dev" # Points to the wrapper

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include both top-level entry-point modules and the friday package
packages = ["friday", "server.py", "agent_friday.py", "main.py"]
